Move MetaStore to filodb.core.store
