+++ coordinator/src/main/scala/filodb.coordinator/NodeCoordinatorActor.scala
@@ -9,8 +9,8 @@ import scala.concurrent.duration._
 
 import filodb.core._
 import filodb.core.Types._
-import filodb.core.metadata.{Column, Dataset, MetaStore, Projection, RichProjection}
-import filodb.core.store.ColumnStore
+import filodb.core.metadata.{Column, Dataset, Projection, RichProjection}
+import filodb.core.store.{ColumnStore, MetaStore}
 import filodb.core.reprojector.Reprojector
 
 /**
